backend/requirements.txt
absl-py==2.1.0
altair==4.2.2
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.2.2
comtypes==1.1.14
configparser==5.3.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.5
decorator==4.4.2
Django==4.2.2
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
dnspython==2.3.0
email-validator==2.0.0.post2
entrypoints==0.4
environs==11.0.0
exceptiongroup==1.1.1
executing==2.0.1
Flask==2.2.3
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.39.4
gast==0.6.0
gitdb==4.0.10
GitPython==3.1.31
google-pasta==0.2.0
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
idna==3.4
imageio==2.29.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
joLib==0.0.1
jsonschema==4.17.3
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.4
libclang==18.1.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.21.2
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mongoengine==0.27.0
moviepy==1.0.3
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.24.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
optree==0.12.1
outcome==1.2.0
packaging==23.1
pandas==1.5.3
parso==0.8.4
Pillow==9.5.0
platformdirs==4.2.2
praw==7.7.0
prawcore==2.3.0
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==12.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
pymongo==4.6.1
Pympler==1.0.1
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pyzmq==26.1.0
requests==2.31.0
rich==13.4.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.9.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.23.1
tenacity==8.2.2
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.14.3
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==4.3
update-checker==0.18.0
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.13
websocket-client==1.5.2
Werkzeug==2.2.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.0
zipp==3.15.0
